refactor(auth): extract parse_scope to break middleware import cycle (#243)
