+++ src/asap/auth/scopes.py
@@ -6,10 +6,11 @@ required scopes on protected routes.
 
 from __future__ import annotations
 
-from typing import Any, Callable, cast
+from typing import Callable, cast
 
 from fastapi import HTTPException, Request
 
+from asap.auth._scope_parse import parse_scope
 from asap.auth.middleware import OAuth2Claims
 
 # Scope constants for ASAP operations
@@ -23,30 +24,13 @@ ERROR_AUTH_REQUIRED = "Authentication required"
 ERROR_INSUFFICIENT_SCOPE = "Insufficient scope"
 
 
-def parse_scope(claim: Any) -> list[str]:
-    """Normalize a scope claim to a list of strings.
-
-    JWT/OAuth2 scope can be a space-separated string (RFC 6749) or a list.
-
-    Args:
-        claim: Raw scope value from JWT claims or introspection response.
-
-    Returns:
-        List of scope strings (empty if claim is None or invalid).
-
-    Example:
-        >>> parse_scope("asap:read asap:execute")
-        ['asap:read', 'asap:execute']
-        >>> parse_scope(["a", "b"])
-        ['a', 'b']
-    """
-    if claim is None:
-        return []
-    if isinstance(claim, list):
-        return [str(s) for s in claim]
-    if isinstance(claim, str):
-        return [s.strip() for s in claim.split() if s.strip()]
-    return []
+__all__ = [
+    "SCOPE_READ",
+    "SCOPE_EXECUTE",
+    "SCOPE_ADMIN",
+    "parse_scope",
+    "require_scope",
+]
 
 
 def require_scope(scope: str) -> Callable[[Request], OAuth2Claims]:
